Init validators stats calculations
